pyproject.toml
[project]
name = "z2g"
version = "0.1.0"
description = "One-way sync Zoho Calendar events to Google"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.189.0",
    "google-auth>=2.48.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
license = { text = "MIT" }
authors = [
    { name = "Jason Hoomani" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Environment :: Console",
    "Topic :: Office/Business :: Scheduling",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "responses>=0.25.8",
]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
z2g = "z2g.cli:main"

